+++ src/ui/templates/Welcome.d.ts
@@ -3,10 +3,10 @@ import { Uri } from "vscode";
 interface WelcomeProps {
     /** Version if the plugin to show to the user. */
     version: string;
-    newWorkspace: Uri;
-    openNewWorkspace: Uri;
     TMCMenuIcon: Uri;
     newTMCMenu: Uri;
+    newTreeView: Uri;
+    actionsExplorer: Uri;
     tmcLogoFile: Uri;
 }