+++ src/api/tmc.ts
@@ -18,10 +18,13 @@ import {
     AuthorizationError,
     ConnectionError,
     ForbiddenError,
+    InvalidTokenError,
+    ObsoleteClientError,
     RuntimeError,
 } from "../errors";
 import { sleep } from "../utils/";
 import { Logger } from "../utils/logger";
+import { showError } from "../window";
 
 import {
     Course,
@@ -554,7 +557,6 @@ export default class TMC {
             return result;
         }
 
-        Logger.debug("reset-exercise", result.val);
         return Ok.EMPTY;
     }
 
@@ -730,13 +732,26 @@ export default class TMC {
                 const traceString = trace.join("\n");
                 Logger.error("TMC Langs errored.", kind, traceString);
                 switch (kind) {
-                    case "not-logged-in":
-                        this._onLogout?.();
-                        return new Err(new AuthorizationError(message, traceString));
                     case "connection-error":
                         return new Err(new ConnectionError(message, traceString));
                     case "forbidden":
                         return new Err(new ForbiddenError(message, traceString));
+                    case "invalid-token":
+                        this._onLogout?.();
+                        showError("Your TMC session has expired, please log in.");
+                        return new Err(new InvalidTokenError(message));
+                    case "not-logged-in":
+                        this._onLogout?.();
+                        return new Err(new AuthorizationError(message, traceString));
+                    case "obsolete-client":
+                        return new Err(
+                            new ObsoleteClientError(
+                                message +
+                                    "\nYoure TMC Extension is out of date, please update it." +
+                                    "\nhttps://code.visualstudio.com/docs/editor/extension-gallery",
+                                traceString,
+                            ),
+                        );
                     default:
                         return new Err(new RuntimeError(message, traceString));
                 }