+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [1.1.0] - 2020-09-21
+
+#### Added
+- Handle logout when token has expired.
+- Log out user if token is invalid.
+- Run TMC Actions for exercises when right clicking exercise folders or files in the explorer view.
+    * This way .ipynb notebooks can be tested and submitted more easily.
+- Buttons to clear new exercises buffer from notification or course card.
+- Inform user if TMC Extension is too old.
+- Access to course details page from the TMC Tree Menu.
+- Refresh button to fetch course and exericse updates in TMC Tree View.
+
+#### Changed
+- Technical improvements to Login page.
+- Technical improvements to TMC TreeView menu.
+- Bumped TMC-langs Rust version to 0.4.0.
+- Exam mode in submissions handled by TMC-langs Rust.
+- Distinguish between disabled courses and access denied for TMC-langs Rust commands.
+- Use course title instead of course name.
+
+#### Security
+- Updated dependencies.
+
+
 ## [1.0.1] - 2020-09-01
 
 #### Fixed